refactoring of equipment repository

+++ Server/DealnetPortal.Domain/NewEquipment.cs
@@ -9,10 +9,10 @@
         public int Id { get; set; }
         public int Quantity { get; set; }
         public string Description { get; set; }
-        public double Cost { get; set; }
-        public double MonthlyCost { get; set; }
+        public decimal Cost { get; set; }
+        public decimal MonthlyCost { get; set; }
         public DateTime? EstimatedInstallationDate { get; set; }
-        public double TotalMonthlyPayment { get; set; }
+        public decimal TotalMonthlyPayment { get; set; }
         public int EquipmentInfoId { get; set; }
         [ForeignKey("EquipmentInfoId")]
         public EquipmentInfo EquipmentInfo { get; set; }
